fix: annoying nasted report filename(s) #1
- Add CSV File model
- Move some util function to model and other util
- Delete filename property from Options model
- Add filename regexp in constants
- Change csv file reader method
